pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sumire"
dynamic = ["version"]
description = 'Scikit-compatible Japanese text vectorizer for CPU-based Japanese natural language processing.'
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
keywords = ["NLP", "Natural Language Processing", "Scikit-learn", "Japanese", "Tokenizer"]
authors = [
    { name = "author", email = "a.k.a.underfirst@mail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    "loguru",
    "inflection",
    "pyknp >=0.4",

    "sudachipy >=0.2",
    "sudachidict_full",

    "mecab-python3",
    "unidic",
    "unidic-lite",
    "ipadic",

    "ginza >=3",
    "ja-ginza >=3",
    "sentencepiece >=0.1.90",
    "scikit-learn >=0.17",

    "gensim",
    "torch",
    "transformers",
    "protobuf",
    "fugashi"
]


[project.urls]
Documentation = "https://github.com/unknown/sumire#readme"
Issues = "https://github.com/unknown/sumire/issues"
Source = "https://github.com/unknown/sumire"

[tool.hatch.version]
path = "sumire/__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
    "sphinx",
    "sphinx-rtd-theme"
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]
cov = [
    "test-cov",
    "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.9", "3.10", "3.11"]

[[tool.hatch.envs.test.matrix]]
python = ["3.9", "3.10", "3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
    "black>=23.1.0",
    "mypy>=1.0.0",
    "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:sumire tests}"
style = [
    "ruff {args:.}",
    "black --check --diff {args:.}",
]
fmt = [
    "black {args:.}",
    "ruff --fix {args:.}",
    "style",
]
all = [
    "style",
    "typing",
]

[tool.black]
target-version = ["py37"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py37"
line-length = 120
select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
    "UP", # pyupgrade
]
ignore = [
    # Allow non-abstract empty methods in abstract base classes
    "B027",
    # Allow boolean positional values in function calls, like `dict.get(... True)`
    "FBT003",
    # Ignore checks for possible passwords
    "S105", "S106", "S107",
    # Ignore complexity
    "C901",
    # Additional Ignore
    "C408", "B904"
]
unfixable = [
    # Don't touch unused imports
    "F401",
]

[tool.ruff.isort]
known-first-party = ["sumire"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]
"sumire/vectorizer/bert_emb.py" = ["E501", "S101"]
"sumire/tokenizer/mecab.py" = ["E501"]
"sumire/tokenizer/spm.py" = ["E501"]
"sumire/utils/download_file.py" = ["PLR2004"]
"sumire/utils/run_sh.py" = ["PLW1510"
]
[tool.coverage.run]
source_pkgs = ["sumire", "tests"]
branch = true
parallel = true
omit = [
    "sumire/__about__.py",
]

[tool.coverage.paths]
sumire = ["sumire", "*/sumire"]
tests = ["tests", "*/sumire/tests"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
